ep10/requirements.txt
claude-agent-sdk>=0.1.0
anthropic>=0.40.0
fastmcp>=2.0.0
